--- src/aosm/azext_aosm/tests/latest/integration_tests/test_aosm_core_vnf_build.py
@@ -1,10 +1,0 @@
-# import unittest
-
-# from azext_aosm.custom import onboard_nfd_build
-# class TestCoreVNFBuild(unittest.TestCase):
-#     # def setUp(self) -> None:
-#     #     # get fixtures?
-
-#     def test_build_with_file_path(self):
-#         file = 'fixtures/core-vnf/core-vnf-input-with-filepath-with-vhd-params.jsonc'
-#         onboard_nfd_build(definition_type="vnf", config_file=file)
